Restructure logging code for better testability
