+++ loglib/logconvert.c
@@ -1,0 +1,106 @@
+/*
+* Author: Christian Huitema
+* Copyright (c) 2019, Private Octopus, Inc.
+* All rights reserved.
+*
+* Permission to use, copy, modify, and distribute this software for any
+* purpose with or without fee is hereby granted, provided that the above
+* copyright notice and this permission notice appear in all copies.
+*
+* THIS SOFTWARE IS PROVIDED BY THE COPYRIGHT HOLDERS AND CONTRIBUTORS "AS IS" AND
+* ANY EXPRESS OR IMPLIED WARRANTIES, INCLUDING, BUT NOT LIMITED TO, THE IMPLIED
+* WARRANTIES OF MERCHANTABILITY AND FITNESS FOR A PARTICULAR PURPOSE ARE
+* DISCLAIMED. IN NO EVENT SHALL Private Octopus, Inc. BE LIABLE FOR ANY
+* DIRECT, INDIRECT, INCIDENTAL, SPECIAL, EXEMPLARY, OR CONSEQUENTIAL DAMAGES
+* (INCLUDING, BUT NOT LIMITED TO, PROCUREMENT OF SUBSTITUTE GOODS OR SERVICES;
+* LOSS OF USE, DATA, OR PROFITS; OR BUSINESS INTERRUPTION) HOWEVER CAUSED AND
+* ON ANY THEORY OF LIABILITY, WHETHER IN CONTRACT, STRICT LIABILITY, OR TORT
+* (INCLUDING NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF THE USE OF THIS
+* SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
+*/
+#include "picoquic_internal.h"
+
+const char * ptype2str(picoquic_packet_type_enum ptype)
+{
+    switch (ptype) {
+    case picoquic_packet_error:
+        return "error";
+    case picoquic_packet_version_negotiation:
+        return "version";
+    case picoquic_packet_initial:
+        return "initial";
+    case picoquic_packet_retry:
+        return "retry";
+    case picoquic_packet_handshake:
+        return "handshake";
+    case picoquic_packet_0rtt_protected:
+        return "0rtt";
+    case picoquic_packet_1rtt_protected:
+        return "1rtt";
+    case picoquic_packet_type_max:
+    default:
+        return "unknown";
+    }
+}
+
+const char * ftype2str(picoquic_frame_type_enum_t ftype)
+{
+    if ((int)ftype >= picoquic_frame_type_stream_range_min &&
+        (int)ftype <= picoquic_frame_type_stream_range_max) {
+        return "stream";
+    }
+
+    switch (ftype) {
+    case picoquic_frame_type_padding:
+        return "padding";
+    case picoquic_frame_type_reset_stream:
+        return "reset_stream";
+    case picoquic_frame_type_connection_close:
+        return "connection_close";
+    case picoquic_frame_type_application_close:
+        return "application_close";
+    case picoquic_frame_type_max_data:
+        return "max_data";
+    case picoquic_frame_type_max_stream_data:
+        return "max_stream_data";
+    case picoquic_frame_type_max_streams_bidir:
+        return "max_streams_bidir";
+    case picoquic_frame_type_max_streams_unidir:
+        return "max_streams_unidir";
+    case picoquic_frame_type_ping:
+        return "ping";
+    case picoquic_frame_type_data_blocked:
+        return "data_blocked";
+    case picoquic_frame_type_stream_data_blocked:
+        return "stream_data_blocked";
+    case picoquic_frame_type_streams_blocked_bidir:
+        return "streams_blocked_bidir";
+    case picoquic_frame_type_streams_blocked_unidir:
+        return "streams_blocked_unidir";
+    case picoquic_frame_type_new_connection_id:
+        return "new_connection_id";
+    case picoquic_frame_type_stop_sending:
+        return "stop_sending";
+    case picoquic_frame_type_ack:
+        return "ack";
+    case picoquic_frame_type_path_challenge:
+        return "path_challenge";
+    case picoquic_frame_type_path_response:
+        return "path_response";
+    case picoquic_frame_type_crypto_hs:
+        return "crypto";
+    case picoquic_frame_type_new_token:
+        return "new_token";
+    case picoquic_frame_type_ack_ecn:
+        return "ack_ecn";
+    case picoquic_frame_type_retire_connection_id:
+        return "retire_connection_id";
+    case picoquic_frame_type_datagram:
+    case picoquic_frame_type_datagram_l:
+    case picoquic_frame_type_datagram_id:
+    case picoquic_frame_type_datagram_id_l:
+        return "datagram";
+    default:
+        return "unknown";
+    }
+}
