+++ loglib/loglib.vcxproj
@@ -140,12 +140,18 @@
   <ItemGroup>
     <ClCompile Include="cidset.c" />
     <ClCompile Include="csv.c" />
+    <ClCompile Include="logconvert.c" />
     <ClCompile Include="logreader.c" />
+    <ClCompile Include="qlog.c" />
+    <ClCompile Include="svg.c" />
   </ItemGroup>
   <ItemGroup>
     <ClInclude Include="cidset.h" />
     <ClInclude Include="csv.h" />
+    <ClInclude Include="logconvert.h" />
     <ClInclude Include="logreader.h" />
+    <ClInclude Include="qlog.h" />
+    <ClInclude Include="svg.h" />
   </ItemGroup>
   <Import Project="$(VCTargetsPath)\Microsoft.Cpp.targets" />
   <ImportGroup Label="ExtensionTargets">
